analyzer-parser/requirements.txt
Flask==3.1.0
pika==1.3.2
python-dotenv==1.0.0
Werkzeug==3.1.0
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
zipfile36==0.1.3